refactor: isolate implementation specific to custom fields
